+++ test/conformance/conformance_test.go
@@ -13,11 +13,9 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	gatewayv1 "sigs.k8s.io/gateway-api/apis/v1"
-	gatewayv1alpha2 "sigs.k8s.io/gateway-api/apis/v1alpha2"
 	gatewayv1beta1 "sigs.k8s.io/gateway-api/apis/v1beta1"
 	"sigs.k8s.io/gateway-api/conformance"
 	conformancev1 "sigs.k8s.io/gateway-api/conformance/apis/v1"
@@ -92,7 +90,7 @@ func runConformance(
 	t *testing.T,
 	gwType gatewayType,
 	kongRouterFlavor consts.RouterFlavor,
-	supportedFeatures sets.Set[features.FeatureName],
+	supportedFeatures []features.FeatureName,
 	cleanupResources bool,
 	skipped []string,
 ) {
@@ -147,11 +145,11 @@ func runConformance(
 		},
 	}
 	opts.Mode = mode
-	opts.ConformanceProfiles = sets.New(
+	opts.ConformanceProfiles = []suite.ConformanceProfileName{
 		suite.GatewayHTTPConformanceProfileName,
 		suite.GatewayGRPCConformanceProfileName,
 		suite.GatewayTLSConformanceProfileName,
-	)
+	}
 	opts.SupportedFeatures = supportedFeatures
 	opts.SkipTests = skipped
 	opts.CleanupBaseResources = cleanupResources
@@ -217,7 +215,7 @@ func waitForConformanceResourcesCleanup(ctx context.Context, cl client.Client, l
 			}
 		}
 
-		var tlsRoutes gatewayv1alpha2.TLSRouteList
+		var tlsRoutes gatewayv1.TLSRouteList
 		if err := cl.List(ctx, &tlsRoutes); err != nil {
 			return false, nil //nolint:nilerr
 		}