+++ test/integration/kic/isolated/udproute_test.go
@@ -154,7 +154,7 @@ func TestUDPRouteEssentials(t *testing.T) {
 					}},
 				},
 			}
-			udpRoute, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Create(ctx, udpRoute, metav1.CreateOptions{})
+			udpRoute, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Create(ctx, udpRoute, metav1.CreateOptions{})
 			assert.NoError(t, err)
 			cleaner.Add(udpRoute)
 			ctx = SetInCtxForT(ctx, t, udpRoute)
@@ -189,10 +189,10 @@ func TestUDPRouteEssentials(t *testing.T) {
 
 			oldParentRefs := udpRoute.Spec.ParentRefs
 			assert.Eventually(t, func() bool {
-				udpRoute, err := gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
+				udpRoute, err := gatewayClient.GatewayV1().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
 				assert.NoError(t, err)
 				udpRoute.Spec.ParentRefs = nil
-				_, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
+				_, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
 				return err == nil
 			}, time.Minute, time.Second)
 
@@ -212,10 +212,10 @@ func TestUDPRouteEssentials(t *testing.T) {
 
 			t.Log("putting the parentRefs back")
 			assert.Eventually(t, func() bool {
-				udpRoute, err := gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
+				udpRoute, err := gatewayClient.GatewayV1().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
 				assert.NoError(t, err)
 				udpRoute.Spec.ParentRefs = oldParentRefs
-				_, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
+				_, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
 				return err == nil
 			}, time.Minute, time.Second)
 
@@ -326,7 +326,7 @@ func TestUDPRouteEssentials(t *testing.T) {
 
 			t.Log("adding an additional backendRef to the UDPRoute")
 			assert.Eventually(t, func() bool {
-				udpRoute, err := gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
+				udpRoute, err := gatewayClient.GatewayV1().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
 				assert.NoError(t, err)
 				udpRoute.Spec.Rules[0].BackendRefs = []gatewayapi.BackendRef{
 					{
@@ -343,7 +343,7 @@ func TestUDPRouteEssentials(t *testing.T) {
 					},
 				}
 
-				_, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
+				_, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
 				return err == nil
 			}, consts.IngressWait, consts.WaitTick)
 
@@ -371,14 +371,14 @@ func TestUDPRouteEssentials(t *testing.T) {
 
 			t.Log("setting the port in ParentRef which does not have a matching listener in Gateway")
 			assert.Eventually(t, func() bool {
-				udpRoute, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
+				udpRoute, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Get(ctx, udpRoute.Name, metav1.GetOptions{})
 				if err != nil {
 					return false
 				}
 				notExistingPort := gatewayapi.PortNumber(81)
 				udpRoute.Spec.ParentRefs[0].Port = &notExistingPort
 				udpRoute.Spec.ParentRefs[0].Name = gatewayapi.ObjectName(service1Name)
-				udpRoute, err = gatewayClient.GatewayV1alpha2().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
+				udpRoute, err = gatewayClient.GatewayV1().UDPRoutes(namespace).Update(ctx, udpRoute, metav1.UpdateOptions{})
 				return err == nil
 			}, time.Minute, time.Second)